add parser and tree.py for refactoring
